Moves RsvpForm to its component folder and renames it to index
